requirements.txt
Django==4.0.5
djangorestframework==3.14.0